+++ R/labels2table.R
@@ -1,5 +1,6 @@
 #' Make table with labels
 #' @param labels A list
+#' @param vars A list
 #' @param moderator A list
 #' @param covar A list
 #' @param serial A logical
@@ -10,14 +11,21 @@
 #' moderator=list(name="skeptic",site=list(c("a","c")))
 #' covar=list(name=c("C1","C2","C3"),site=list(c("M","Y"),c("M","Y"),c("M","Y")))
 #' labels=list(X="X",M=c("M1","M2","M3"),Y="Y")
+#' labels=list(X="X",M=c("M1","M2"),Y="Y")
 #' moderator=list();serial=FALSE;eq=NULL
-#' labels2table(labels,serial=TRUE)
+#' labels2table(labels)
+#' labels2table(labels,serial=FALSE)
 #' labels2table(labels,covar=covar)
 #' labels2table(labels,moderator=moderator)
-labels2table=function(labels=labels,moderator=list(),covar=NULL,serial=FALSE,eq=NULL){
+labels2table=function(labels=labels,vars=list(),moderator=list(),covar=NULL,serial=TRUE,eq=NULL){
     # moderator=list();serial=FALSE;eq=NULL
+    # labels=list(X="X",M=c("M1","M2"),Y="Y")
+    # vars=list(name=list(c("W","Z")),site=list("a"),arr.pos=list(c(0.5)))
+    # covar=NULL;vars=list()
+
     if(is.null(eq)) {
-      eq=multipleMediation(labels=labels,moderator=moderator,covar=covar,mode=1,serial=serial)
+      eq=multipleMediation(labels=labels,vars=vars,moderator=moderator,covar=covar,mode=1,serial=serial)
+
     }
     eq
     labels